自动测试埋点/requirements.txt
Flask>=3.0.0,<4.0.0
pymysql>=1.1.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
